topicTracker/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pika==1.3.2
pydantic==2.10.4
pydantic_core==2.27.2
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
tavily-python==0.5.0
tiktoken==0.8.0
typing_extensions==4.12.2
urllib3==2.3.0
